throwing out old code, splitting of regioncount_annotator
