+++ gffquant/annotation/__init__.py
@@ -3,6 +3,7 @@
 """ module docstring """
 
 # from .count_annotator import GeneCountAnnotator, RegionCountAnnotator
-from .count_annotator import RegionCountAnnotator
+from .count_annotator import CountAnnotator
 from .count_writer import CountWriter
 from .genecount_annotator import GeneCountAnnotator
+from .regioncount_annotator import RegionCountAnnotator
